Moved config app into the core app
