+++ openslides/agenda/apps.py
@@ -12,7 +12,7 @@ class AgendaAppConfig(AppConfig):
 
         # Import all required stuff.
         from django.db.models.signals import pre_delete
-        from openslides.config.signals import config_signal
+        from openslides.core.signals import config_signal
         from openslides.utils.rest_api import router
         from .signals import setup_agenda_config, listen_to_related_object_delete_signal
         from .views import ItemViewSet
