--- openslides/config/exceptions.py
@@ -1,9 +1,0 @@
-from openslides.utils.exceptions import OpenSlidesError
-
-
-class ConfigError(OpenSlidesError):
-    pass
-
-
-class ConfigNotFound(ConfigError):
-    pass
