--- openslides/config/signals.py
@@ -1,4 +1,0 @@
-from django.dispatch import Signal
-
-config_signal = Signal(providing_args=[])
-"""Signal to get all config tabs from all apps."""
