+++ openslides/core/apps.py
@@ -12,11 +12,16 @@ class CoreAppConfig(AppConfig):
 
         # Import all required stuff.
         from django.db.models import signals
-        from openslides.config.signals import config_signal
+        from openslides.core.signals import config_signal
         from openslides.utils.autoupdate import inform_changed_data_receiver
         from openslides.utils.rest_api import router
         from .signals import setup_general_config
-        from .views import CustomSlideViewSet, ProjectorViewSet, TagViewSet
+        from .views import (
+            ConfigViewSet,
+            CustomSlideViewSet,
+            ProjectorViewSet,
+            TagViewSet,
+        )
 
         # Connect signals.
         config_signal.connect(setup_general_config, dispatch_uid='setup_general_config')
@@ -25,6 +30,7 @@ class CoreAppConfig(AppConfig):
         router.register('core/projector', ProjectorViewSet)
         router.register('core/customslide', CustomSlideViewSet)
         router.register('core/tag', TagViewSet)
+        router.register('core/config', ConfigViewSet, 'config')
 
         # Update data when any model of any installed app is saved or deleted.
         # TODO: Test if the m2m_changed signal is also needed.
