+++ openslides/motions/models.py
@@ -6,7 +6,7 @@ from django.utils.translation import ugettext as _
 from django.utils.translation import ugettext_lazy, ugettext_noop
 from jsonfield import JSONField
 
-from openslides.config.api import config
+from openslides.core.config import config
 from openslides.core.models import Tag
 from openslides.mediafiles.models import Mediafile
 from openslides.poll.models import (
@@ -17,7 +17,7 @@ from openslides.poll.models import (
 )
 from openslides.projector.models import SlideMixin
 from openslides.users.models import User
-from openslides.utils.rest_api import RESTModelMixin
+from openslides.utils.models import RESTModelMixin
 
 from .exceptions import WorkflowError
