+++ openslides/utils/models.py
@@ -1,3 +1,4 @@
+from django.core.urlresolvers import reverse
 from django.db import models
 
 
@@ -14,3 +15,27 @@ class MinMaxIntegerField(models.IntegerField):
         defaults = {'min_value': self.min_value, 'max_value': self.max_value}
         defaults.update(kwargs)
         return super(MinMaxIntegerField, self).formfield(**defaults)
+
+
+class RESTModelMixin:
+    """
+    Mixin for django models which are used in our rest api.
+    """
+
+    def get_root_rest_element(self):
+        """
+        Returns the root rest instance.
+
+        Uses self as default.
+        """
+        return self
+
+    def get_root_rest_url(self):
+        """
+        Returns the detail url of the root model of this object.
+        """
+        # Gets the default url-name in the same way as django rest framework
+        # does in relations.HyperlinkedModelSerializer
+        root_instance = self.get_root_rest_element()
+        rest_url = '%s-detail' % type(root_instance)._meta.object_name.lower()
+        return reverse(rest_url, args=[str(root_instance.pk)])
