+++ openslides/utils/rest_api.py
@@ -1,7 +1,6 @@
 import re
 from urllib.parse import urlparse
 
-from django.core.urlresolvers import reverse
 from rest_framework.decorators import detail_route  # noqa
 from rest_framework.decorators import list_route  # noqa
 from rest_framework.metadata import SimpleMetadata  # noqa
@@ -31,30 +30,6 @@ from rest_framework.viewsets import (  # noqa
 from .exceptions import OpenSlidesError
 
 router = DefaultRouter()
-
-
-class RESTModelMixin:
-    """
-    Mixin for django models which are used in our rest api.
-    """
-
-    def get_root_rest_element(self):
-        """
-        Returns the root rest instance.
-
-        Uses self as default.
-        """
-        return self
-
-    def get_root_rest_url(self):
-        """
-        Returns the detail url of the root model of this object.
-        """
-        # Gets the default url-name in the same way as django rest framework
-        # does in relations.HyperlinkedModelSerializer
-        root_instance = self.get_root_rest_element()
-        rest_url = '%s-detail' % type(root_instance)._meta.object_name.lower()
-        return reverse(rest_url, args=[str(root_instance.pk)])
 
 
 class ModelViewSet(_ModelViewSet):
